para que no desborde hay que declarar las sumaMax y sumaMin Decimal o Double

+++ HRank_SumaMax_Min/HRank_SumaMax_Min/Program.cs
@@ -23,19 +23,21 @@ class Result
 
     public static void miniMaxSum(List<int> arr)
     {
-        int sumaMax = 0, sumaMin = 0;
+        decimal sumaMax = 0, sumaMin = 0;
+        
         arr.Sort();     //Ordenom de menor a mayor
         
-        for(int i = 0; i < 4; i++)
+        for (int i = 0; i < 4; i++)
         {
             sumaMin = sumaMin + arr[i];
-            sumaMax = sumaMax + arr[arr.Count -1 - i];
+
+            sumaMax = sumaMax + arr[arr.Count - 1 - i]; 
         }
         //arr.Reverse();      //Invierto la cade`na para ordenar de mayor a menor
 
         //for (int i = 0; i < 4; i++)
         //{
-        //    sumaMax = sumaMax + arr[i];
+        //    sumaMax = sumaMax + arr[i]
         //}
 
         Console.Write(sumaMin + " " + sumaMax);
